LMS_CAPSTONE_2025-main/Search/requirements.txt
pip
flask
motor
fastapi
uvicorn[standard]
google-cloud-firestore
firebase-admin
python-dotenv
pydantic
pymongo
python-jose
passlib
requests
pytest
python-multipart
pytz
